Move checkApiKey to `_utils` folder
Lets follow the existing convention to use underscore-folder for components, utils and such.

Also if-else-switch for easier readability.
